Fix chain breakage due to disruption of message handling
