+++ src/main/java/net/earthcomputer/clientcommands/mixin/MixinClientPlayNetworkHandler.java
@@ -86,72 +86,4 @@ public class MixinClientPlayNetworkHandler {
             ServerBrandManager.setServerBrand(this.client.player.getServerBrand());
         }
     }
-
-    @Inject(method = "onChatMessage", at = @At("HEAD"), cancellable = true)
-    public void onC2CPacket(ChatMessageS2CPacket packet, CallbackInfo ci) {
-        Optional<MessageType.Parameters> optionalParameters = packet.getParameters(this.registryManager);
-        if (optionalParameters.isEmpty()) {
-            return;
-        }
-        Text content = optionalParameters.get().applyChatDecoration(packet.message().getContent());
-        handleIfPacket(content, ci);
-    }
-
-    @Inject(method = "onGameMessage", at = @At("HEAD"), cancellable = true)
-    public void onC2CPacket(GameMessageS2CPacket packet, CallbackInfo ci) {
-        Text content = packet.content();
-        handleIfPacket(content, ci);
-    }
-
-    private void handleIfPacket(Text content, CallbackInfo ci) {
-        String string = content.getString();
-        int index = string.indexOf("CCENC:");
-        if (index == -1) {
-            return;
-        }
-        String packetString = string.substring(index + 6);
-        if (!TempRules.acceptC2CPackets) {
-            if (OutgoingPacketFilter.removeIfContains(packetString)) {
-                this.client.inGameHud.getChatHud().addMessage(Text.translatable("ccpacket.sentC2CPacket"));
-            } else {
-                this.client.inGameHud.getChatHud().addMessage(Text.translatable("ccpacket.receivedC2CPacket").styled(s -> s.withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, content))));
-            }
-            ci.cancel();
-            return;
-        }
-        if (OutgoingPacketFilter.removeIfContains(packetString)) {
-            ci.cancel();
-            return;
-        }
-        if (handleC2CPacket(packetString)) {
-            ci.cancel();
-        } else {
-            this.client.inGameHud.getChatHud().addMessage(Text.translatable("ccpacket.malformedPacket"));
-        }
-    }
-
-    private static boolean handleC2CPacket(String content) {
-        byte[] encrypted = ConversionHelper.BaseUTF8.fromUnicode(content);
-        encrypted = Arrays.copyOf(encrypted, 256);
-        Optional<PrivateKey> key = ((IProfileKeys) MinecraftClient.getInstance().getProfileKeys()).getPrivateKey();
-        if (key.isEmpty()) {
-            return false;
-        }
-        byte[] decrypted = ConversionHelper.RsaEcb.decrypt(encrypted, key.get());
-        if (decrypted == null) {
-            return false;
-        }
-        byte[] uncompressed = ConversionHelper.Gzip.uncompress(decrypted);
-        StringBuf buf = new StringBuf(new String(uncompressed, StandardCharsets.UTF_8));
-        int id = buf.readInt();
-        CCPacket ccPacket = CCPacketHandler.createPacket(id, buf);
-        if (ccPacket == null) {
-            return false;
-        }
-        if (buf.getRemainingLength() > 0) {
-            return false;
-        }
-        ccPacket.apply(CCNetworkHandler.getInstance());
-        return true;
-    }
 }
