+++ src/main/resources/mixins.clientcommands.json
@@ -29,6 +29,7 @@
     "MixinAxeHoeAndShovelItem",
     "MixinBowAndTridentItem",
     "MixinCartographyTableContainer",
+    "MixinChatHud",
     "MixinClientCommandSource",
     "MixinClientPlayerEntity",
     "MixinClientPlayerInteractionManager",
